requirements.txt
amqp==5.3.1
anyio==4.12.0
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
backports.tarfile==1.2.0
beautifulsoup4==4.14.3
billiard==4.2.4
blivet==3.12.1
boto3==1.42.1
botocore==1.42.1
Brlapi==0.8.7
build==1.3.0
CacheControl==0.14.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cockpit @ file:///builddir/build/BUILD/cockpit-356-build/cockpit-356/tmp/wheel/cockpit-356-py3-none-any.whl#sha256=5a84a26ba9e7f611c00c6efb3a3e675039514318856e1355bd01b72f1d86f01c
colorama==0.4.6
crashtest==0.4.1
crypt_r==3.13.1
cryptography==46.0.5
cupshelpers==1.0
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
distlib==0.4.0
distro==1.9.0
dj-database-url==3.1.1
Django==5.2.9
django-channels==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
django-ratelimit==4.1.0
django-redis==6.0.0
django-resized==1.0.3
django-rest-passwordreset==1.5.0
django-sslserver==0.22
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnf==4.24.0
drf-spectacular==0.29.0
drf-yasg==1.21.11
dulwich==0.24.10
exceptiongroup==1.3.1
fastjsonschema==2.21.2
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.20.0
findpython==0.7.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
installer==0.7.0
iso639==0.1.4
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
kombu==5.6.1
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
louis==3.33.0
lxml==6.0.1
more-itertools==10.8.0
msgpack==1.1.2
nftables==0.1
oauthlib==3.3.1
olefile==0.47
packaging==25.0
paystackapi==2.1.3
pbs-installer==2025.11.20
perf==0.1
pexpect==4.9.0
phonenumbers==9.0.19
pid==3.0.4
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.5.0
ply==3.11
poetry==2.2.1
poetry-core==2.2.1
productmd==1.50
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.54.5
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pyparted==3.13.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-discovery==1.1.0
python-dotenv==1.2.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pytz==2025.2
pyudev==0.24.3
pywin32-ctypes==0.2.3
pyxdg==0.28
PyYAML==6.0.3
pyynl @ file:///builddir/build/BUILD/kernel-6.18.13-build/kernel-6.18.13/linux-6.18.13-200.fc43.x86_64/tools/net/ynl
RapidFuzz==3.14.3
razorpay==2.0.0
redis==7.1.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
rpm==6.0.1
s3transfer==0.16.0
SecretStorage==3.5.0
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
shellingham==1.5.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.11.0
soupsieve==2.8
sqlparse==0.5.4
stripe==14.0.1
systemd-python==235
tomlkit==0.13.3
trove-classifiers==2025.12.1.14
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
whitenoise==6.11.0
xkbregistry==0.3
zipp==3.23.0
zstandard==0.25.0
